refactor. shared util 함수 분리(useQueryParams, highlightText)
